BUG: no implicit casting
On some versions of numpy, implicit cast raises an error:

TypeError: Cannot cast array data from dtype('uint64') to dtype('int64')
according to the rule 'safe'

+++ nipy/labs/mask.py
@@ -38,7 +38,7 @@ def largest_cc(mask):
         raise ValueError('No non-zero values: no connected components')
     if label_nb == 1:
         return mask.astype(np.bool)
-    label_count = np.bincount(labels.ravel())
+    label_count = np.bincount(labels.ravel().astype(np.int))
     # discard 0 the 0 label
     label_count[0] = 0
     return labels == label_count.argmax()
